lte: socket: Fix bug that mbedtls_net_connect() fails
Fix bug that mbedtls_net_connect() fails depending on
the registered IP address when CONFIG_LTE_NET_MBEDTLS enabled.

When the registered address is only IPv6, it failed to connect to the server.
Because the altcom_getaddrinfo() used in mbedtls_net_connect() always
returned the IPv4 server address.
Fixed to set an appropriate address family based on the registered
IP address when ALTCOM_AF_UNSPEC is set in hints.ai_family.

refs #11996